Allow Repository without /site folder

+++ recipe/trellis.php
@@ -10,7 +10,12 @@ namespace Deployer;
 // Remove trellis, so /site directory becomes the web root
 desc( 'Removes Trellis (checked in in Git) and moves /site content up to root release path' );
 task( 'trellis:remove', function () {
-    run( 'mv {{release_path}}/site/* {{release_path}}' );
-    run( 'rm -rf {{release_path}}/site' );
-    run( 'rm -rf {{release_path}}/trellis' );
+    if ( test('[ -d {{release_path}}/site ]') ) {
+        run( 'mv {{release_path}}/site/* {{release_path}}' );
+        run( 'rm -rf {{release_path}}/site' );
+        run( 'rm -rf {{release_path}}/trellis' );
+    } else {
+        writeln("<comment>No /site folder found. Skipping.</comment>");
+        return;
+    }
 } );
